Moved functions out of event gather pipeline to more relevant files
