+++ cdp_backend/database/functions.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import logging
+from datetime import datetime
 from typing import Optional
 
 import fireo
@@ -9,7 +10,9 @@ from fireo.models import Model
 from prefect import task
 
 from ..database import exceptions
+from ..database import models as db_models
 from ..database.validators import get_model_uniqueness
+from ..pipeline import ingestion_models
 from ..pipeline.ingestion_models import IngestionModel
 
 ###############################################################################
@@ -142,3 +145,78 @@ def upload_db_model_task(
     return upload_db_model(
         db_model=db_model, ingestion_model=ingestion_model, creds_file=creds_file
     )
+
+
+@task
+def create_body_from_ingestion_model(body: ingestion_models.Body) -> db_models.Body:
+    db_body = db_models.Body()
+
+    # Required fields
+    db_body.name = body.name
+    db_body.is_active = body.is_active
+    if body.start_datetime is None:
+        db_body.start_datetime = datetime.utcnow()
+    else:
+        db_body.start_datetime = body.start_datetime
+
+    # Optional fields
+    if body.end_datetime:
+        db_body.end_datetime = body.end_datetime
+
+    if body.description:
+        db_body.description = body.description
+
+    if body.external_source_id:
+        db_body.external_source_id = body.external_source_id
+
+    return db_body
+
+
+@task
+def create_event_from_ingestion_model(
+    event: ingestion_models.EventIngestionModel, body_ref: db_models.Body
+) -> db_models.Event:
+    db_event = db_models.Event()
+
+    # Required fields
+    db_event.body_ref = body_ref
+
+    # Assume event datetime is the date of earliest session
+    db_event.event_datetime = min(
+        [session.session_datetime for session in event.sessions]
+    )
+
+    # TODO add optional fields
+
+    return db_event
+
+
+@task
+def create_session_from_ingestion_model(
+    session: ingestion_models.Session, event_ref: db_models.Event
+) -> db_models.Session:
+    db_session = db_models.Session()
+
+    # Required fields
+    db_session.event_ref = event_ref
+    db_session.session_datetime = session.session_datetime
+    db_session.video_uri = session.video_uri
+    db_session.session_index = session.session_index
+
+    # Optional fields
+    if session.caption_uri:
+        db_session.caption_uri = session.caption_uri
+
+    if session.external_source_id:
+        db_session.external_source_id = session.external_source_id
+
+    return db_session
+
+
+@task
+def create_file(name: str, uri: str) -> db_models.File:
+    db_file = db_models.File()
+    db_file.name = name
+    db_file.uri = uri
+
+    return db_file
